add view vocab term
