+++ src/lib/app/components.ts
@@ -5,6 +5,7 @@ import SpaceDelete from '$lib/ui/SpaceDelete.svelte';
 import SpaceInput from '$lib/ui/SpaceInput.svelte';
 import type {SvelteComponent} from 'svelte';
 import {contextmenuComponents} from '$lib/app/contextmenu/components';
+import {viewComponents} from '$lib/vocab/view/components';
 
 // The collection of components that can be dynamically mounted by the app.
 
@@ -18,4 +19,5 @@ export const components: {[key: string]: typeof SvelteComponent} = {
 	SpaceDelete,
 	SpaceInput,
 	...contextmenuComponents,
+	...viewComponents,
 };
