+++ src/lib/vocab/view/components.ts
@@ -1,6 +1,6 @@
 import type {SvelteComponent} from 'svelte';
 
-import type {ViewType} from '$lib/vocab/space/space';
+import type {ViewType} from '$lib/vocab/view/view';
 import Home from '$lib/ui/Home.svelte';
 import Room from '$lib/ui/Room.svelte';
 import Board from '$lib/ui/Board.svelte';
@@ -9,7 +9,7 @@ import Notes from '$lib/ui/Notes.svelte';
 import Iframe from '$lib/ui/Iframe.svelte';
 import Voice from '$lib/ui/Voice.svelte';
 
-export const spaceViews: Record<ViewType, typeof SvelteComponent> = {
+export const viewComponents: Record<ViewType, typeof SvelteComponent> = {
 	Home,
 	Room,
 	Board,
